requirements.txt
djangoajax==3.2
Django==3.2
channels == 3.0.4
python-dateutil==2.8.2